pulling models into individual modules and fixing some imports for inter-module dependencies
