+++ pep/models.py
@@ -58,11 +58,6 @@ import itertools
 import logging
 from operator import itemgetter
 import os as _os
-import sys
-if sys.version_info < (3, 0):
-    from urlparse import urlparse
-else:
-    from urllib.parse import urlparse
 import warnings
 
 import pandas as _pd
@@ -78,13 +73,9 @@ from .utils import \
 
 # TODO: decide if we want to denote functions for export.
 __functions__ = []
-__classes__ = ["AttributeDict", "PipelineInterface", "Project",
-               "ProtocolInterface", "ProtocolMapper", "Sample"]
-__all__ = __functions__ + __classes__
 
 
 MAX_PROJECT_SAMPLES_REPR = 12
-ATTRDICT_METADATA = {"_force_nulls": False, "_attribute_identity": False}
 
 _LOGGER = logging.getLogger(__name__)
 if not logging.getLogger().handlers:
@@ -121,236 +112,6 @@ def check_sheet(sample_file, dtype=str):
 
 
 
-def copy(obj):
-    def copy(self):
-        """
-        Copy self to a new object.
-        """
-        from copy import deepcopy
-
-        return deepcopy(self)
-    obj.copy = copy
-    return obj
-
-
-
-def fetch_samples(proj, inclusion=None, exclusion=None):
-    """
-    Collect samples of particular protocol(s).
-
-    Protocols can't be both positively selected for and negatively
-    selected against. That is, it makes no sense and is not allowed to
-    specify both inclusion and exclusion protocols. On the other hand, if
-    neither is provided, all of the Project's Samples are returned.
-    If inclusion is specified, Samples without a protocol will be excluded,
-    but if exclusion is specified, protocol-less Samples will be included.
-
-    :param Project proj: the Project with Samples to fetch
-    :param Iterable[str] | str inclusion: protocol(s) of interest;
-        if specified, a Sample must
-    :param Iterable[str] | str exclusion: protocol(s) to include
-    :return list[Sample]: Collection of this Project's samples with
-        protocol that either matches one of those in inclusion, or either
-        lacks a protocol or does not match one of those in exclusion
-    :raise TypeError: if both inclusion and exclusion protocols are
-        specified; TypeError since it's basically providing two arguments
-        when only one is accepted, so remain consistent with vanilla Python2
-    """
-
-    # Intersection between inclusion and exclusion is nonsense user error.
-    if inclusion and exclusion:
-        raise TypeError("Specify only inclusion or exclusion protocols, "
-                         "not both.")
-
-    if not inclusion and not exclusion:
-        # Simple; keep all samples.  In this case, this function simply
-        # offers a list rather than an iterator.
-        return list(proj.samples)
-
-    # Ensure that we're working with sets.
-    def make_set(items):
-        if isinstance(items, str):
-            items = [items]
-        return {alpha_cased(i) for i in items}
-
-    # Use the attr check here rather than exception block in case the
-    # hypothetical AttributeError would occur in alpha_cased; we want such
-    # an exception to arise, not to catch it as if the Sample lacks "protocol"
-    if not inclusion:
-        # Loose; keep all samples not in the exclusion.
-        def keep(s):
-            return not hasattr(s, "protocol") or \
-                   alpha_cased(s.protocol) not in make_set(exclusion)
-    else:
-        # Strict; keep only samples in the inclusion.
-        def keep(s):
-            return hasattr(s, "protocol") and \
-                   alpha_cased(s.protocol) in make_set(inclusion)
-
-    return list(filter(keep, proj.samples))
-
-
-
-def include_in_repr(attr, klazz):
-    """
-    Determine whether to include attribute in an object's text representation.
-
-    :param str attr: attribute to include/exclude from object's representation
-    :param str | type klazz: name of type or type itself of which the object
-        to be represented is an instance
-    :return bool: whether to include attribute in an object's
-        text representation
-    """
-    # TODO: try to leverage the class hierarchy to determine these exclusions.
-    ad_metadata = list(ATTRDICT_METADATA.keys())
-    exclusions_by_class = {
-            AttributeDict.__name__: ad_metadata,
-            Project.__name__: ["_samples", "merge_table", "sheet",
-                               "interfaces_by_protocol"] + ad_metadata,
-            Sample.__name__: ["sheet", "prj", "merged_cols"] + ad_metadata}
-    classname = klazz.__name__ if isinstance(klazz, type) else klazz
-    return attr not in exclusions_by_class.get(classname, [])
-
-
-
-def is_url(maybe_url):
-    """
-    Determine whether a path is a URL.
-
-    :param str maybe_url: path to investigate as URL
-    :return bool: whether path appears to be a URL
-    """
-    return urlparse(maybe_url).scheme != ""
-
-
-
-def merge_sample(sample, merge_table, data_sources=None, derived_columns=None):
-    """
-    Use merge table data to augment/modify Sample.
-
-    :param Sample sample: sample to modify via merge table data
-    :param merge_table: data with which to alter Sample
-    :param Mapping data_sources: collection of named paths to data locations,
-        optional
-    :param Iterable[str] derived_columns: names of columns for which
-        corresponding Sample attribute's value is data-derived, optional
-    :return Set[str]: names of columns that were merged
-    """
-
-    merged_attrs = {}
-
-    if merge_table is None:
-        _LOGGER.log(5, "No data for sample merge, skipping")
-        return merged_attrs
-
-    if SAMPLE_NAME_COLNAME not in merge_table.columns:
-        raise KeyError(
-            "Merge table requires a column named '{}'.".
-                format(SAMPLE_NAME_COLNAME))
-
-    _LOGGER.debug("Merging Sample with data sources: {}".
-                  format(data_sources))
-    
-    # Hash derived columns for faster lookup in case of many samples/columns.
-    derived_columns = set(derived_columns or [])
-    _LOGGER.debug("Merging Sample with derived columns: {}".
-                  format(derived_columns))
-
-    sample_name = getattr(sample, SAMPLE_NAME_COLNAME)
-    sample_indexer = merge_table[SAMPLE_NAME_COLNAME] == sample_name
-    this_sample_rows = merge_table[sample_indexer]
-    if len(this_sample_rows) == 0:
-        _LOGGER.debug("No merge rows for sample '%s', skipping", sample.name)
-        return merged_attrs
-    _LOGGER.log(5, "%d rows to merge", len(this_sample_rows))
-    _LOGGER.log(5, "Merge rows dict: {}".format(this_sample_rows.to_dict()))
-
-    # For each row in the merge table of this sample:
-    # 1) populate any derived columns
-    # 2) derived columns --> space-delimited strings
-    # 3) update the sample values with the merge table
-    # Keep track of merged cols,
-    # so we don't re-derive them later.
-    merged_attrs = {key: "" for key in this_sample_rows.columns}
-    
-    for _, row in this_sample_rows.iterrows():
-        rowdata = row.to_dict()
-
-        # Iterate over column names to avoid Python3 RuntimeError for
-        # during-iteration change of dictionary size.
-        for attr_name in this_sample_rows.columns:
-            if attr_name == SAMPLE_NAME_COLNAME or \
-                            attr_name not in derived_columns:
-                _LOGGER.log(5, "Skipping merger of attribute '%s'", attr_name)
-                continue
-
-            attr_value = rowdata[attr_name]
-
-            # Initialize key in parent dict.
-            col_key = attr_name + COL_KEY_SUFFIX
-            merged_attrs[col_key] = ""
-            rowdata[col_key] = attr_value
-            data_src_path = sample.locate_data_source(
-                    data_sources, attr_name, source_key=rowdata[attr_name],
-                    extra_vars=rowdata)  # 1)
-            rowdata[attr_name] = data_src_path
-
-        _LOGGER.log(5, "Adding derived columns")
-        
-        for attr in derived_columns:
-            
-            # Skip over any attributes that the sample lacks or that are
-            # covered by the data from the current (row's) data.
-            if not hasattr(sample, attr) or attr in rowdata:
-                _LOGGER.log(5, "Skipping column: '%s'", attr)
-                continue
-            
-            # Map key to sample's value for the attribute given by column name.
-            col_key = attr + COL_KEY_SUFFIX
-            rowdata[col_key] = getattr(sample, attr)
-            # Map the col/attr name itself to the populated data source 
-            # template string.
-            rowdata[attr] = sample.locate_data_source(
-                    data_sources, attr, source_key=getattr(sample, attr),
-                    extra_vars=rowdata)
-
-        # TODO: this (below) is where we could maintain grouped values
-        # TODO (cont.): as a collection and defer the true merger.
-
-        # Since we are now jamming multiple (merged) entries into a single
-        # attribute on a Sample, we have to join the individual items into a
-        # space-delimited string and then use that value as the Sample
-        # attribute. The intended use case for this sort of merge is for
-        # multiple data source paths associated with a single Sample, hence
-        # the choice of space-delimited string as the joined-/merged-entry
-        # format--it's what's most amenable to use in building up an argument
-        # string for a pipeline command.
-        for attname, attval in rowdata.items():
-            if attname == SAMPLE_NAME_COLNAME or not attval:
-                _LOGGER.log(5, "Skipping KV: {}={}".format(attname, attval))
-                continue
-            _LOGGER.log(5, "merge: sample '%s'; '%s'='%s'",
-                        str(sample.name), str(attname), str(attval))
-            if attname not in merged_attrs:
-                new_attval = str(attval).rstrip()
-            else:
-                new_attval = "{} {}".format(merged_attrs[attname], str(attval)).strip()
-            merged_attrs[attname] = new_attval  # 2)
-            _LOGGER.log(5, "Stored '%s' as value for '%s' in merged_attrs",
-                        new_attval, attname)
-
-    # If present, remove sample name from the data with which to update sample.
-    merged_attrs.pop(SAMPLE_NAME_COLNAME, None)
-
-    _LOGGER.log(5, "Updating Sample {}: {}".format(sample.name, merged_attrs))
-    sample.update(merged_attrs)  # 3)
-    sample.merged_cols = merged_attrs
-    sample.merged = True
-
-    return sample
-
-
-
 def process_pipeline_interfaces(pipeline_interface_locations):
     """
     Create a ProtocolInterface for each pipeline location given.
@@ -384,2065 +145,10 @@ SubmissionBundle = namedtuple(
 SUBMISSION_BUNDLE_PIPELINE_KEY_INDEX = 2
 
 
-
-@copy
-class Paths(object):
-    """ A class to hold paths as attributes. """
-
-    def __getitem__(self, key):
-        """
-        Provides dict-style access to attributes
-        """
-        return getattr(self, key)
-
-    def __iter__(self):
-        """
-        Iteration is over the paths themselves.
-        
-        Note that this implementation constrains the assignments to be 
-        non-nested. That is, the value for any attribute attached to the 
-        instance should be a path.
-        
-        """
-        return iter(self.__dict__.values())
-
-    def __repr__(self):
-        return "Paths object."
-
-
-
-class ProjectContext(object):
-    """ Wrap a Project to provide protocol-specific Sample selection. """
-
-    def __init__(self, prj, include_protocols=None, exclude_protocols=None):
-        """ Project and what to include/exclude defines the context. """
-        self.prj = prj
-        self.include = include_protocols
-        self.exclude = exclude_protocols
-
-    def __getattr__(self, item):
-        """ Samples are context-specific; other requests are handled
-        locally or dispatched to Project. """
-        if item == "samples":
-            return fetch_samples(
-                self.prj, inclusion=self.include, exclusion=self.exclude)
-        if item in ["prj", "include", "exclude"]:
-            # Attributes requests that this context/wrapper handles
-            return self.__dict__[item]
-        else:
-            # Dispatch attribute request to Project.
-            return getattr(self.prj, item)
-
-    def __getitem__(self, item):
-        return self.prj[item]
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, *args):
-        pass
-
-
-
 class IFilteredRepr(object):
     def __repr__(self):
         return
 
-
-
-
-@copy
-class AttributeDict(MutableMapping):
-    """
-    A class to convert a nested mapping(s) into an object(s) with key-values
-    using object syntax (attr_dict.attribute) instead of getitem syntax 
-    (attr_dict["key"]). This class recursively sets mappings to objects, 
-    facilitating attribute traversal (e.g., attr_dict.attr.attr).
-    """
-
-    def __init__(self, entries=None,
-                 _force_nulls=False, _attribute_identity=False):
-        """
-        Establish a logger for this instance, set initial entries,
-        and determine behavior with regard to null values and behavior
-        for attribute requests.
-
-        :param collections.Iterable | collections.Mapping entries: collection
-            of key-value pairs, initial data for this mapping
-        :param bool _force_nulls: whether to allow a null value to overwrite
-            an existing non-null value
-        :param bool _attribute_identity: whether to return attribute name
-            requested rather than exception when unset attribute/key is queried
-        """
-        # Null value can squash non-null?
-        self.__dict__["_force_nulls"] = _force_nulls
-        # Return requested attribute name if not set?
-        self.__dict__["_attribute_identity"] = _attribute_identity
-        self.add_entries(entries)
-
-
-    def add_entries(self, entries):
-        """
-        Update this `AttributeDict` with provided key-value pairs.
-
-        :param Iterable[(object, object)] | Mapping | pandas.Series entries:
-            collection of pairs of keys and values
-        """
-        if entries is None:
-            return
-        # Permit mapping-likes and iterables/generators of pairs.
-        if callable(entries):
-            entries = entries()
-        elif isinstance(entries, _pd.Series):
-            entries = entries.to_dict()
-        try:
-            entries_iter = entries.items()
-        except AttributeError:
-            entries_iter = entries
-        # Assume we now have pairs; allow corner cases to fail hard here.
-        for key, value in entries_iter:
-            self.__setitem__(key, value)
-
-
-    def __setattr__(self, key, value):
-        self.__setitem__(key, value)
-
-
-    def __getattr__(self, item, default=None):
-        """
-        Fetch the value associated with the provided identifier.
-
-        :param int | str item: identifier for value to fetch
-        :return object: whatever value corresponds to the requested key/item
-        :raises AttributeError: if the requested item has not been set,
-            no default value is provided, and this instance is not configured
-            to return the requested key/item itself when it's missing; also,
-            if the requested item is unmapped and appears to be protected,
-            i.e. by flanking double underscores, then raise AttributeError
-            anyway. More specifically, respect attribute naming that appears
-            to be indicative of the intent of protection.
-        """
-        try:
-            return super(AttributeDict, self).__getattribute__(item)
-        except (AttributeError, TypeError):
-            # Handle potential property and non-string failures.
-            pass
-        try:
-            # Fundamentally, this is still a mapping;
-            # route object notation access pattern accordingly.
-            # Ideally, the requested item maps to a value.
-            return self.__dict__[item]
-        except KeyError:
-            # If not, arbitrage and cope accordingly.
-            if item.startswith("__") and item.endswith("__"):
-                # Some libraries use exception for protected attribute
-                # access as a control flow mechanism.
-                error_reason = "Protected-looking attribute: {}".format(item)
-                raise AttributeError(error_reason)
-            if default is not None:
-                # For compatibility with ordinary getattr() invocation, allow
-                # caller the ability to provide a default value.
-                return default
-            if self.__dict__.setdefault("_attribute_identity", False):
-                # Check if we should return the attribute name as the value.
-                return item
-            # Throw up our hands in despair and resort to exception behavior.
-            raise AttributeError(item)
-
-
-    def __setitem__(self, key, value):
-        """
-        This is the key to making this a unique data type. Flag set at
-        time of construction determines whether it's possible for a null
-        value to squash a non-null value. The combination of that flag and
-        one indicating whether request for value for unset attribute should
-        return the attribute name itself determines if any attribute/key
-        may be set to a null value.
-
-        :param str key: name of the key/attribute for which to establish value
-        :param object value: value to which set the given key; if the value is
-            a mapping-like object, other keys' values may be combined.
-        :raises _MetadataOperationException: if attempt is made
-            to set value for privileged metadata key
-        """
-        if isinstance(value, Mapping):
-            try:
-                # Combine AttributeDict instances.
-                self.__dict__[key].add_entries(value)
-            except (AttributeError, KeyError):
-                # Create new AttributeDict, replacing previous value.
-                self.__dict__[key] = AttributeDict(value)
-        elif value is not None or \
-                key not in self.__dict__ or self.__dict__["_force_nulls"]:
-            self.__dict__[key] = value
-
-
-    def __getitem__(self, item):
-        try:
-            # Ability to return requested item name itself is delegated.
-            return self.__getattr__(item)
-        except AttributeError:
-            # Requested item is unknown, but request was made via
-            # __getitem__ syntax, not attribute-access syntax.
-            raise KeyError(item)
-
-    def __delitem__(self, item):
-        if item in ATTRDICT_METADATA:
-            raise _MetadataOperationException(self, item)
-        try:
-            del self.__dict__[item]
-        except KeyError:
-            _LOGGER.debug("No item {} to delete".format(item))
-
-    def __eq__(self, other):
-        try:
-            # Ensure target itself and any values are AttributeDict.
-            other = AttributeDict(other)
-        except Exception:
-            return False
-        if len(self) != len(other):
-            # Ensure we don't have to worry about other containing self.
-            return False
-        for k, v in self.items():
-            try:
-                if v != other[k]:
-                    return False
-            except KeyError:
-                return  False
-        return True
-
-    def __ne__(self, other):
-        return not self == other
-
-    def __iter__(self):
-        return iter([k for k in self.__dict__.keys()
-                     if k not in ATTRDICT_METADATA])
-
-    def __len__(self):
-        return sum(1 for _ in iter(self))
-
-    def __repr__(self):
-        return repr({k: v for k, v in self.__dict__.items()
-                    if include_in_repr(k, klazz=self.__class__)})
-
-    def __str__(self):
-        return "{}: {}".format(self.__class__.__name__, repr(self))
-
-
-
-@copy
-class Project(AttributeDict):
-    """
-    A class to model a Project.
-
-    :param config_file: Project config file (YAML).
-    :type config_file: str
-    :param subproject: Subproject to use within configuration file, optional
-    :type subproject: str
-    :param default_compute: Configuration file (YAML) for 
-        default compute settings.
-    :type default_compute: str
-    :param dry: If dry mode is activated, no directories 
-        will be created upon project instantiation.
-    :type dry: bool
-    :param permissive: Whether a error should be thrown if 
-        a sample input file(s) do not exist or cannot be open.
-    :type permissive: bool
-    :param file_checks: Whether sample input files should be checked 
-        for their  attributes (read type, read length) 
-        if this is not set in sample metadata.
-    :type file_checks: bool
-    :param compute_env_file: Environment configuration YAML file specifying
-        compute settings.
-    :type compute_env_file: str
-    :param no_environment_exception: type of exception to raise if environment
-        settings can't be established, optional; if null (the default),
-        a warning message will be logged, and no exception will be raised.
-    :type no_environment_exception: type
-    :param no_compute_exception: type of exception to raise if compute
-        settings can't be established, optional; if null (the default),
-        a warning message will be logged, and no exception will be raised.
-    :type no_compute_exception: type
-    :param defer_sample_construction: whether to wait to build this Project's
-        Sample objects until they're needed, optional; by default, the basic
-        Sample is created during Project construction
-    :type defer_sample_construction: bool
-
-
-    :Example:
-
-    .. code-block:: python
-
-        from models import Project
-        prj = Project("config.yaml")
-    
-    """
-
-    DERIVED_COLUMNS_DEFAULT = [DATA_SOURCE_COLNAME]
-
-
-    def __init__(self, config_file, subproject=None,
-                 default_compute=None, dry=False,
-                 permissive=True, file_checks=False, compute_env_file=None,
-                 no_environment_exception=None, no_compute_exception=None,
-                 defer_sample_construction=False):
-
-        _LOGGER.debug("Creating %s from file: '%s'",
-                          self.__class__.__name__, config_file)
-        super(Project, self).__init__()
-
-        # Initialize local, serial compute as default (no cluster submission)
-        # Start with default environment settings.
-        _LOGGER.debug("Establishing default environment settings")
-        self.environment, self.environment_file = None, None
-
-        try:
-            self.update_environment(
-                    default_compute or self.default_compute_envfile)
-        except Exception as e:
-            _LOGGER.error("Can't load environment config file '%s'",
-                          str(default_compute))
-            _LOGGER.error(str(type(e).__name__) + str(e))
-        
-        self._handle_missing_env_attrs(
-                default_compute, when_missing=no_environment_exception)
-
-        # Load settings from environment yaml for local compute infrastructure.
-        if compute_env_file:
-            _LOGGER.debug("Updating environment settings based on file '%s'",
-                          compute_env_file)
-            self.update_environment(compute_env_file)
-
-        else:
-            _LOGGER.info("Using default {envvar}. You may set environment "
-                         "variable {envvar} to configure environment "
-                         "settings.".format(envvar=self.compute_env_var))
-
-        # Initialize default compute settings.
-        _LOGGER.debug("Establishing project compute settings")
-        self.compute = None
-        self.set_compute(DEFAULT_COMPUTE_RESOURCES_NAME)
-
-        # Either warn or raise exception if the compute is null.
-        if self.compute is None:
-            message = "Failed to establish project compute settings"
-            if no_compute_exception:
-                no_compute_exception(message)
-            else:
-                _LOGGER.warn(message)
-        else:
-            _LOGGER.debug("Compute: %s", str(self.compute))
-
-        # Optional behavioral parameters
-        self.permissive = permissive
-        self.file_checks = file_checks
-
-        # Include the path to the config file.
-        self.config_file = _os.path.abspath(config_file)
-
-        # Parse config file
-        _LOGGER.debug("Parsing %s config file", self.__class__.__name__)
-        if subproject:
-            _LOGGER.info("Using subproject: '{}'".format(subproject))
-        self.parse_config_file(subproject)
-
-        # Ensure data_sources is at least set if it wasn't parsed.
-        self.setdefault("data_sources", None)
-
-        self.name = self.infer_name(self.config_file)
-
-        # Set project's directory structure
-        if not dry:
-            _LOGGER.debug("Ensuring project directories exist")
-            self.make_project_dirs()
-
-        # Establish derived columns.
-        try:
-            # Do not duplicate derived column names.
-            self.derived_columns.extend(
-                    [colname for colname in self.DERIVED_COLUMNS_DEFAULT
-                     if colname not in self.derived_columns])
-        except AttributeError:
-            self.derived_columns = self.DERIVED_COLUMNS_DEFAULT
-
-        self.finalize_pipelines_directory()
-
-        # SampleSheet creation populates project's samples, adds the
-        # sheet itself, and adds any derived columns.
-        _LOGGER.debug("Processing {} pipeline location(s): {}".
-                      format(len(self.metadata.pipelines_dir),
-                             self.metadata.pipelines_dir))
-        self.interfaces_by_protocol = \
-                process_pipeline_interfaces(self.metadata.pipelines_dir)
-
-        path_anns_file = self.metadata.sample_annotation
-        _LOGGER.debug("Reading sample annotations sheet: '%s'", path_anns_file)
-        try:
-            _LOGGER.info("Setting sample sheet from file '%s'", path_anns_file)
-            self.sheet = check_sheet(path_anns_file)
-        except IOError:
-            _LOGGER.error("Alleged annotations file doesn't exist: '%s'",
-                          path_anns_file)
-            anns_folder_path = _os.path.dirname(path_anns_file)
-            try:
-                annotations_file_folder_contents = \
-                        _os.listdir(anns_folder_path)
-            except OSError:
-                _LOGGER.error("Annotations file folder doesn't exist either: "
-                              "'%s'", anns_folder_path)
-            else:
-                _LOGGER.error("Annotations file folder's contents: {}".
-                              format(annotations_file_folder_contents))
-            raise
-
-        self.merge_table = None
-
-        # Basic sample maker will handle name uniqueness check.
-        if defer_sample_construction:
-            self._samples = None
-        else:
-            self._set_basic_samples()
-
-
-    def __repr__(self):
-        """ Representation in interpreter. """
-        samples_message = "{} (from '{}')".\
-                format(self.__class__.__name__, self.config_file)
-        try:
-            num_samples = len(self._samples)
-        except AttributeError:
-            pass
-        else:
-            samples_message += " with {} sample(s)".format(num_samples)
-            if num_samples <= MAX_PROJECT_SAMPLES_REPR:
-                samples_message += ": {}".format(repr(self._samples))
-        meta_text = super(Project, self).__repr__()
-        return "{} -- {}".format(samples_message, meta_text)
-
-
-    @property
-    def compute_env_var(self):
-        """
-        Environment variable through which to access compute settings.
-
-        :return str: name of the environment variable to pointing to
-            compute settings
-        """
-        return COMPUTE_SETTINGS_VARNAME
-
-
-    @property
-    def constants(self):
-        """
-        Return key-value pairs of pan-Sample constants for this Project.
-
-        :return Mapping: collection of KV pairs, each representing a pairing
-            of attribute name and attribute value
-        """
-        return self._constants
-
-
-    @property
-    def default_compute_envfile(self):
-        """ Path to default compute environment settings file. """
-        return _os.path.join(
-                self.templates_folder, "default_compute_settings.yaml")
-
-
-    @property
-    def num_samples(self):
-        """ Number of samples available in this Project. """
-        return sum(1 for _ in self.sample_names)
-
-
-    @property
-    def output_dir(self):
-        """
-        Directory in which to place results and submissions folders.
-
-        By default, assume that the project's configuration file specifies
-        an output directory, and that this is therefore available within
-        the project metadata. If that assumption does not hold, though,
-        consider the folder in which the project configuration file lives
-        to be the project's output directory.
-
-        :return str: path to the project's output directory, either as
-            specified in the configuration file or the folder that contains
-            the project's configuration file.
-        """
-        try:
-            return self.metadata.output_dir
-        except AttributeError:
-            return _os.path.dirname(self.config_file)
-
-
-    @property
-    def project_folders(self):
-        """
-        Names of folders to nest within a project output directory.
-
-        :return Iterable[str]: names of output-nested folders
-        """
-        return ["results_subdir", "submission_subdir"]
-
-
-    @property
-    def protocols(self):
-        """
-        Determine this Project's unique protocol names.
-
-        :return Set[str]: collection of this Project's unique protocol names
-        """
-        protos = set()
-        for s in self.samples:
-            try:
-                protos.add(s.protocol)
-            except AttributeError:
-                _LOGGER.debug("Sample '%s' lacks protocol", s.sample_name)
-        return protos
-
-
-    @property
-    def required_metadata(self):
-        """
-        Names of metadata fields that must be present for a valid project.
-        
-        Make a base project as unconstrained as possible by requiring no 
-        specific metadata attributes. It's likely that some common-sense 
-        requirements may arise in domain-specific client applications, in 
-        which case this can be redefined in a subclass.
-        
-        :return Iterable[str]: names of metadata fields required by a project
-        """
-        return []
-
-
-    @property
-    def sample_names(self):
-        """ Names of samples of which this Project is aware. """
-        return iter(self.sheet[SAMPLE_NAME_COLNAME])
-
-
-    @property
-    def samples(self):
-        """
-        Generic/base Sample instance for each of this Project's samples.
-
-        :return Iterable[Sample]: Sample instance for each
-            of this Project's samples
-        """
-        if self._samples is None:
-            _LOGGER.debug("Building basic sample object(s) for %s",
-                          self.__class__.__name__)
-            self._set_basic_samples()
-        return self._samples
-
-
-    @property
-    def templates_folder(self):
-        """
-        Path to folder with default submission templates.
-
-        :return str: path to folder with default submission templates
-        """
-        return _os.path.join(_os.path.dirname(__file__), "submit_templates")
-
-
-    @staticmethod
-    def infer_name(path_config_file):
-        """
-        Infer project name based on location of configuration file.
-        
-        Provide the project with a name, taken to be the name of the folder 
-        in which its configuration file lives.
-        
-        :param str path_config_file: path to the project's configuration file.
-        :return str: name of the configuration file's folder, to name project.
-        """
-        config_dirpath = _os.path.dirname(path_config_file)
-        _, config_folder = _os.path.split(config_dirpath)
-        return config_folder
-
-
-    def build_sheet(self, *protocols):
-        """
-        Create all Sample object for this project for the given protocol(s).
-
-        :return pandas.core.frame.DataFrame: DataFrame with from base version
-            of each of this Project's samples, for indicated protocol(s) if
-            given, else all of this Project's samples
-        """
-        # Use all protocols if none are explicitly specified.
-        protocols = {alpha_cased(p) for p in (protocols or self.protocols)}
-        include_samples = []
-        for s in self.samples:
-            try:
-                proto = s.protocol
-            except AttributeError:
-                include_samples.append(s)
-                continue
-            check_proto = alpha_cased(proto)
-            if check_proto in protocols:
-                include_samples.append(s)
-            else:
-                _LOGGER.debug("Sample skipped due to protocol ('%s')", proto)
-        return _pd.DataFrame(include_samples)
-
-
-    def build_submission_bundles(self, protocol, priority=True):
-        """
-        Create pipelines to submit for each sample of a particular protocol.
-
-        With the argument (flag) to the priority parameter, there's control
-        over whether to submit pipeline(s) from only one of the project's
-        known pipeline locations with a match for the protocol, or whether to
-        submit pipelines created from all locations with a match for the
-        protocol.
-        
-        :param str protocol: name of the protocol/library for which to
-            create pipeline(s)
-        :param bool priority: to only submit pipeline(s) from the first of the
-            pipelines location(s) (indicated in the project config file) that
-            has a match for the given protocol; optional, default True
-        :return Iterable[(PipelineInterface, type, str, str)]:
-        :raises AssertionError: if there's a failure in the attempt to
-            partition an interface's pipeline scripts into disjoint subsets of
-            those already mapped and those not yet mapped
-        """
-
-        if not priority:
-            raise NotImplementedError(
-                "Currently, only prioritized protocol mapping is supported "
-                "(i.e., pipeline interfaces collection is a prioritized list, "
-                "so only the first interface with a protocol match is used.)")
-
-        # Pull out the collection of interfaces (potentially one from each of
-        # the locations indicated in the project configuration file) as a
-        # sort of pool of information about possible ways in which to submit
-        # pipeline(s) for sample(s) of the indicated protocol.
-        try:
-            protocol_interfaces = \
-                    self.interfaces_by_protocol[protocol]
-        except KeyError:
-            # Messaging can be done by the caller.
-            return []
-
-        job_submission_bundles = []
-        pipeline_keys_used = set()
-        _LOGGER.debug("Building pipelines for {} PIs...".
-                      format(len(protocol_interfaces)))
-
-        bundle_by_strict_pipe_key = {}
-
-        for proto_iface in protocol_interfaces:
-            # "Break"-like mechanism for short-circuiting if we care only
-            # about the highest-priority match for pipeline submission.
-            # That is, if the intent is to submit pipeline(s) from a single
-            # location for each sample of the given protocol, we can stop
-            # searching the pool of pipeline interface information once we've
-            # found a match for the protocol.
-            if priority and len(job_submission_bundles) > 0:
-                return job_submission_bundles[0]
-
-            this_protocol_pipelines = proto_iface.fetch_pipelines(protocol)
-            if not this_protocol_pipelines:
-                continue
-            
-            # TODO: update once dependency-encoding logic is in place.
-            # The proposed dependency-encoding format uses a semicolon
-            # between pipelines for which the dependency relationship is
-            # serial. For now, simply treat those as multiple independent
-            # pipelines by replacing the semicolon with a comma, which is the
-            # way in which multiple independent pipelines for a single protocol
-            # are represented in the mapping declaration.
-            pipeline_keys = \
-                    this_protocol_pipelines.replace(";", ",")\
-                                           .strip(" ()\n")\
-                                           .split(",")
-            # These cleaned pipeline keys are what's used to resolve the path
-            # to the pipeline to run.
-            pipeline_keys = [pk.strip() for pk in pipeline_keys]
-
-            # Skip over pipelines already mapped by another location.
-            already_mapped, new_scripts = \
-                    partition(pipeline_keys,
-                              partial(_is_member, items=pipeline_keys_used))
-            pipeline_keys_used |= set(pipeline_keys)
-
-            # Attempt to validate that partition yielded disjoint subsets.
-            try:
-                disjoint_partition_violation = \
-                        set(already_mapped) & set(new_scripts)
-            except TypeError:
-                _LOGGER.debug("Unable to hash partitions for validation")
-            else:
-                assert not disjoint_partition_violation, \
-                        "Partitioning {} with membership in {} as " \
-                        "predicate produced intersection: {}".format(
-                        pipeline_keys, pipeline_keys_used, 
-                        disjoint_partition_violation)
-
-            if len(already_mapped) > 0:
-                _LOGGER.debug("Skipping {} already-mapped script name(s): {}".
-                              format(len(already_mapped), already_mapped))
-            _LOGGER.debug("{} new scripts for protocol {} from "
-                          "pipeline(s) location '{}': {}".
-                          format(len(new_scripts), protocol,
-                                 proto_iface.source, new_scripts))
-
-            pl_iface = proto_iface.pipe_iface
-
-            # For each pipeline script to which this protocol will pertain,
-            # create the new jobs/submission bundles.
-            new_jobs = []
-            for pipeline_key in new_scripts:
-                # Determine how to reference the pipeline and where it is.
-                strict_pipe_key, full_pipe_path, full_pipe_path_with_flags = \
-                        proto_iface.finalize_pipeline_key_and_paths(
-                                pipeline_key)
-
-                # Skip and warn about nonexistent alleged pipeline path.
-                if not (_os.path.exists(full_pipe_path) or
-                        is_command_callable(full_pipe_path)):
-                    _LOGGER.warn("Missing pipeline script: '%s'",
-                                 full_pipe_path)
-                    continue
-
-                # Determine which interface and Sample subtype to use.
-                sample_subtype = \
-                        proto_iface.fetch_sample_subtype(
-                                protocol, strict_pipe_key, full_pipe_path)
-
-                # Package the pipeline's interface, subtype, command, and key.
-                submission_bundle = SubmissionBundle(
-                        pl_iface, sample_subtype, strict_pipe_key,
-                        full_pipe_path_with_flags)
-
-                # Enforce bundle unqiueness for each strict pipeline key.
-                maybe_new_bundle = (full_pipe_path_with_flags,
-                                    sample_subtype, pl_iface)
-                old_bundle = bundle_by_strict_pipe_key.setdefault(
-                        strict_pipe_key, maybe_new_bundle)
-                if old_bundle != maybe_new_bundle:
-                    errmsg = "Strict pipeline key '{}' maps to more than " \
-                            "one combination of pipeline script + flags, " \
-                            "sample subtype, and pipeline interface. " \
-                             "'{}'\n{}".format(
-                            strict_pipe_key, maybe_new_bundle, old_bundle)
-                    raise ValueError(errmsg)
-
-                # Add this bundle to the collection of ones relevant for the
-                # current ProtocolInterface.
-                new_jobs.append(submission_bundle)
-
-            job_submission_bundles.append(new_jobs)
-
-        # Repeat logic check of short-circuit conditional to account for
-        # edge case in which it's satisfied during the final iteration.
-        if priority and len(job_submission_bundles) > 1:
-            return job_submission_bundles[0]
-        else:
-            return list(itertools.chain(*job_submission_bundles))
-
-
-    def _check_unique_samples(self):
-        """ Handle scenario in which sample names are not unique. """
-        # Defining this here but then calling out to the repeats counter has
-        # a couple of advantages. We get an unbound, isolated method (the
-        # Project-external repeat sample name counter), but we can still
-        # do this check from the sample builder, yet have it be override-able.
-        repeats = {name: n for name, n in Counter(
-                s.name for s in self._samples).items() if n > 1}
-        if repeats:
-            histogram_text = "\n".join(
-                    "{}: {}".format(name, n) for name, n in repeats.items())
-            _LOGGER.warn("Non-unique sample names:\n{}".format(histogram_text))
-
-
-    def finalize_pipelines_directory(self, pipe_path=""):
-        """
-        Finalize the establishment of a path to this project's pipelines.
-
-        With the passed argument, override anything already set.
-        Otherwise, prefer path provided in this project's config, then
-        local pipelines folder, then a location set in project environment.
-
-        :param str pipe_path: (absolute) path to pipelines
-        :raises PipelinesException: if (prioritized) search in attempt to
-            confirm or set pipelines directory failed
-        :raises TypeError: if pipeline(s) path(s) argument is provided and
-            can't be interpreted as a single path or as a flat collection
-            of path(s)
-        """
-
-        # TODO: check for local pipelines or those from environment.
-
-        # Pass pipeline(s) dirpath(s) or use one already set.
-        if not pipe_path:
-            try:
-                # TODO: beware of AttributeDict with _attribute_identity = True
-                #  here, as that may return 'pipelines_dir' name itself.
-                pipe_path = self.metadata.pipelines_dir
-            except AttributeError:
-                pipe_path = []
-
-        # Ensure we're working with a flattened list.
-        if isinstance(pipe_path, str):
-            pipe_path = [pipe_path]
-        elif isinstance(pipe_path, Iterable) and \
-                not isinstance(pipe_path, Mapping):
-            pipe_path = list(pipe_path)
-        else:
-            _LOGGER.debug("Got {} as pipelines path(s) ({})".
-                          format(pipe_path, type(pipe_path)))
-            pipe_path = []
-
-        self.metadata.pipelines_dir = pipe_path
-
-
-    def get_arg_string(self, pipeline_name):
-        """
-        For this project, given a pipeline, return an argument string
-        specified in the project config file.
-        """
-
-        def make_optarg_text(opt, arg):
-            """ Transform flag/option into CLI-ready text version. """
-            if arg:
-                try:
-                    arg = _os.path.expandvars(arg)
-                except TypeError:
-                    # Rely on direct string formatting of arg.
-                    pass
-                return "{} {}".format(opt, arg)
-            else:
-                return opt
-
-        def create_argtext(name):
-            """ Create command-line argstring text from config section. """
-            try:
-                optargs = getattr(self.pipeline_args, name)
-            except AttributeError:
-                return ""
-            # NS using __dict__ will add in the metadata from AttrDict (doh!)
-            _LOGGER.debug("optargs.items(): {}".format(optargs.items()))
-            optargs_texts = [make_optarg_text(opt, arg)
-                             for opt, arg in optargs.items()]
-            _LOGGER.debug("optargs_texts: {}".format(optargs_texts))
-            # TODO: may need to fix some spacing issues here.
-            return " ".join(optargs_texts)
-
-        default_argtext = create_argtext(DEFAULT_COMPUTE_RESOURCES_NAME)
-        _LOGGER.debug("Creating additional argstring text for pipeline '%s'",
-                      pipeline_name)
-        pipeline_argtext = create_argtext(pipeline_name)
-
-        if not pipeline_argtext:
-            # The project config may not have an entry for this pipeline;
-            # no problem! There are no pipeline-specific args. Return text
-            # from default arguments, whether empty or not.
-            return default_argtext
-        elif default_argtext:
-            # Non-empty pipeline-specific and default argtext
-            return " ".join([default_argtext, pipeline_argtext])
-        else:
-            # No default argtext, but non-empty pipeline-specific argtext
-            return pipeline_argtext
-
-
-    def make_project_dirs(self):
-        """
-        Creates project directory structure if it doesn't exist.
-        """
-        for folder_name in self.project_folders:
-            folder_path = self.metadata[folder_name]
-            _LOGGER.debug("Ensuring project dir exists: '%s'", folder_path)
-            if not _os.path.exists(folder_path):
-                _LOGGER.debug("Attempting to create project folder: '%s'",
-                              folder_path)
-                try:
-                    _os.makedirs(folder_path)
-                except OSError as e:
-                    _LOGGER.warn("Could not create project folder: '%s'",
-                                 str(e))
-
-
-    def _set_basic_samples(self):
-        """ Build the base Sample objects from the annotations sheet data. """
-
-        # This should be executed just once, establishing the Project's
-        # base Sample objects if they don't already exist.
-        if hasattr(self.metadata, "merge_table"):
-            if self.merge_table is None:
-                if self.metadata.merge_table and \
-                        _os.path.isfile(self.metadata.merge_table):
-                    _LOGGER.info("Reading merge table: %s",
-                                 self.metadata.merge_table)
-                    self.merge_table = _pd.read_table(
-                        self.metadata.merge_table,
-                        sep=None, engine="python")
-                    _LOGGER.debug("Merge table shape: {}".
-                                  format(self.merge_table.shape))
-                else:
-                    _LOGGER.debug(
-                        "Alleged path to merge table data is not a "
-                        "file: '%s'", self.metadata.merge_table)
-            else:
-                _LOGGER.debug("Already parsed merge table")
-        else:
-            _LOGGER.debug("No merge table")
-
-        # Set samples and handle non-unique names situation.
-        self._samples = self._prep_samples()
-        self._check_unique_samples()
-
-
-    def _prep_samples(self):
-        """
-        Merge this Project's Sample object and set file paths.
-
-        :return list[Sample]: collection of this Project's Sample objects
-        """
-
-        samples = []
-
-        for _, row in self.sheet.iterrows():
-            sample = Sample(row.dropna(), prj=self)
-
-            # Add values that are constant across this Project's samples.
-            sample = add_project_sample_constants(sample, self)
-
-            # TODO: use implied_columns in 0.8.
-            sample.set_genome(self.get("genomes"))
-            sample.set_transcriptome(self.get("transcriptomes"))
-
-            _LOGGER.debug("Merging sample '%s'", sample.name)
-            merge_sample(sample, self.merge_table,
-                         self.data_sources, self.derived_columns)
-            _LOGGER.debug("Setting sample file paths")
-            sample.set_file_paths(self)
-            # Hack for backwards-compatibility
-            # Pipelines should now use `data_source`)
-            _LOGGER.debug("Setting sample data path")
-            try:
-                sample.data_path = sample.data_source
-            except AttributeError:
-                _LOGGER.log(5, "Sample '%s' lacks data source; skipping "
-                              "data path assignment", sample.sample_name)
-            else:
-                _LOGGER.log(5, "Path to sample data: '%s'", sample.data_source)
-            samples.append(sample)
-
-        return samples
-
-
-    def parse_config_file(self, subproject=None):
-        """
-        Parse provided yaml config file and check required fields exist.
-        
-        :raises KeyError: if config file lacks required section(s)
-        """
-
-        _LOGGER.debug("Setting %s data from '%s'",
-                      self.__class__.__name__, self.config_file)
-        with open(self.config_file, 'r') as conf_file:
-            config = yaml.safe_load(conf_file)
-
-        _LOGGER.debug("{} config data: {}".format(
-                self.__class__.__name__, config))
-
-        # Parse yaml into the project's attributes.
-        _LOGGER.debug("Adding attributes for {}: {}".format(
-                self.__class__.__name__, config.keys()))
-        _LOGGER.debug("Config metadata: {}".format(config["metadata"]))
-        self.add_entries(config)
-        _LOGGER.debug("{} now has {} keys: {}".format(
-                self.__class__.__name__, len(self.keys()), self.keys()))
-
-        # Overwrite any config entries with entries in the subproject.
-        if "subprojects" in config and subproject:
-            _LOGGER.debug("Adding entries for subproject '{}'".
-                          format(subproject))
-            subproj_updates = config['subprojects'][subproject]
-            _LOGGER.debug("Updating with: {}".format(subproj_updates))
-            self.add_entries(subproj_updates)
-        else:
-            _LOGGER.debug("No subproject")
-
-        # In looper 0.4, for simplicity the paths section was eliminated.
-        # For backwards compatibility, mirror the paths section into metadata.
-        if "paths" in config:
-            _LOGGER.warn(
-                "Paths section in project config is deprecated. "
-                "Please move all paths attributes to metadata section. "
-                "This option will be removed in future versions.")
-            self.metadata.add_entries(self.paths)
-            _LOGGER.debug("Metadata: %s", str(self.metadata))
-            delattr(self, "paths")
-
-        # In looper 0.6, we added pipeline_interfaces to metadata
-        # For backwards compatibility, merge it with pipelines_dir
-
-        if "metadata" in config:
-            if "pipelines_dir" in self.metadata:
-                _LOGGER.warning("Looper v0.6 suggests "
-                    "switching from pipelines_dir to "
-                    "pipeline_interfaces. See docs for details: "
-                    "https://pepkit.github.io/docs/home/")
-            if "pipeline_interfaces" in self.metadata:
-                if "pipelines_dir" in self.metadata:
-                    raise AttributeError(
-                            "You defined both 'pipeline_interfaces' and "
-                            "'pipelines_dir'. Please remove your "
-                            "'pipelines_dir' definition.")
-                else:
-                    self.metadata.pipelines_dir = \
-                            self.metadata.pipeline_interfaces
-                _LOGGER.debug("Adding pipeline_interfaces to "
-                    "pipelines_dir. New value: {}".
-                    format(self.metadata.pipelines_dir))
-
-        self._constants = config.get("constants", dict())
-
-        # Ensure required absolute paths are present and absolute.
-        for var in self.required_metadata:
-            if var not in self.metadata:
-                raise ValueError("Missing required metadata item: '%s'")
-            setattr(self.metadata, var,
-                    _os.path.expandvars(getattr(self.metadata, var)))
-
-        _LOGGER.debug("{} metadata: {}".format(self.__class__.__name__,
-                                               self.metadata))
-
-        # These are optional because there are defaults
-        config_vars = {
-            # Defaults = {"variable": "default"}, relative to output_dir.
-            "results_subdir": "results_pipeline",
-            "submission_subdir": "submission"
-        }
-
-        for key, value in config_vars.items():
-            if hasattr(self.metadata, key):
-                if not _os.path.isabs(getattr(self.metadata, key)):
-                    setattr(self.metadata, key,
-                            _os.path.join(self.output_dir,
-                                          getattr(self.metadata, key)))
-            else:
-                outpath = _os.path.join(self.output_dir, value)
-                setattr(self.metadata, key, outpath)
-
-        # Variables which are relative to the config file
-        # All variables in these sections should be relative to project config.
-        relative_sections = ["metadata", "pipeline_config"]
-
-        _LOGGER.debug("Parsing relative sections")
-        for sect in relative_sections:
-            if not hasattr(self, sect):
-                _LOGGER.log(5, "%s lacks relative section '%s', skipping",
-                            self.__class__.__name__, sect)
-                continue
-            relative_vars = getattr(self, sect)
-            if not relative_vars:
-                _LOGGER.log(5, "No relative variables, continuing")
-                continue
-            for var in relative_vars.keys():
-                if not hasattr(relative_vars, var) or \
-                                getattr(relative_vars, var) is None:
-                    continue
-
-                relpath = getattr(relative_vars, var)
-                _LOGGER.debug("Ensuring absolute path(s) for '%s'", var)
-                # Parsed from YAML, so small space of possible datatypes.
-                if isinstance(relpath, list):
-                    absolute = [self._ensure_absolute(maybe_relpath)
-                                for maybe_relpath in relpath]
-                else:
-                    absolute = self._ensure_absolute(relpath)
-                _LOGGER.debug("Setting '%s' to '%s'", var, absolute)
-                setattr(relative_vars, var, absolute)
-
-        # Project config may have made compute.submission_template relative.
-        # Make sure it's absolute.
-        if self.compute is None:
-            _LOGGER.log(5, "No compute, no submission template")
-        elif not _os.path.isabs(self.compute.submission_template):
-            # Relative to environment config file.
-            self.compute.submission_template = _os.path.join(
-                    _os.path.dirname(self.environment_file),
-                    self.compute.submission_template)
-
-        # Required variables check
-        if not hasattr(self.metadata, SAMPLE_ANNOTATIONS_KEY):
-            raise _MissingMetadataException(
-                    missing_section=SAMPLE_ANNOTATIONS_KEY, 
-                    path_config_file=self.config_file)
-
-
-    def set_compute(self, setting):
-        """
-        Set the compute attributes according to the
-        specified settings in the environment file.
-
-        :param str setting:	name for non-resource compute bundle, the name of
-            a subsection in an environment configuration file
-        :return bool: success flag for attempt to establish compute settings
-        """
-
-        # Hope that environment & environment compute are present.
-        if setting and self.environment and "compute" in self.environment:
-            # Augment compute, creating it if needed.
-            if self.compute is None:
-                _LOGGER.debug("Creating Project compute")
-                self.compute = AttributeDict()
-                _LOGGER.debug("Adding entries for setting '%s'", setting)
-            self.compute.add_entries(self.environment.compute[setting])
-
-            # Ensure submission template is absolute.
-            if not _os.path.isabs(self.compute.submission_template):
-                try:
-                    self.compute.submission_template = _os.path.join(
-                            _os.path.dirname(self.environment_file),
-                            self.compute.submission_template)
-                except AttributeError as e:
-                    # Environment and environment compute should at least have been
-                    # set as null-valued attributes, so execution here is an error.
-                    _LOGGER.error(str(e))
-                    # Compute settings have been established.
-                else:
-                    return True
-        else:
-            # Scenario in which environment and environment compute are
-            # both present--but don't evaluate to True--is fairly harmless.
-            _LOGGER.debug("Environment = {}".format(self.environment))
-
-        return False
-
-
-    def set_project_permissions(self):
-        """
-        Make the project's public_html folder executable.
-        """
-        try:
-            _os.chmod(self.trackhubs.trackhub_dir, 0o0755)
-        except OSError:
-            # This currently does not fail now
-            # ("cannot change folder's mode: %s" % d)
-            pass
-
-
-    def update_environment(self, env_settings_file):
-        """
-        Parse data from environment configuration file.
-
-        :param str env_settings_file: path to file with 
-            new environment configuration data
-        """
-
-        with open(env_settings_file, 'r') as f:
-            _LOGGER.info("Loading %s: %s",
-                         self.compute_env_var, env_settings_file)
-            env_settings = yaml.load(f)
-            _LOGGER.debug("Parsed environment settings: %s",
-                          str(env_settings))
-
-            # Any compute.submission_template variables should be made
-            # absolute, relative to current environment settings file.
-            y = env_settings["compute"]
-            for key, value in y.items():
-                if type(y[key]) is dict:
-                    for key2, value2 in y[key].items():
-                        if key2 == "submission_template":
-                            if not _os.path.isabs(y[key][key2]):
-                                y[key][key2] = _os.path.join(
-                                        _os.path.dirname(env_settings_file),
-                                        y[key][key2])
-
-            env_settings["compute"] = y
-            if self.environment is None:
-                self.environment = AttributeDict(env_settings)
-            else:
-                self.environment.add_entries(env_settings)
-
-        self.environment_file = env_settings_file
-
-
-    def _ensure_absolute(self, maybe_relpath):
-        """ Ensure that a possibly relative path is absolute. """
-        _LOGGER.log(5, "Ensuring absolute: '%s'", maybe_relpath)
-        if _os.path.isabs(maybe_relpath) or is_url(maybe_relpath):
-            _LOGGER.log(5, "Already absolute")
-            return maybe_relpath
-        # Maybe we have env vars that make the path absolute?
-        expanded = _os.path.expanduser(_os.path.expandvars(maybe_relpath))
-        _LOGGER.log(5, "Expanded: '%s'", expanded)
-        if _os.path.isabs(expanded):
-            _LOGGER.log(5, "Expanded is absolute")
-            return expanded
-        _LOGGER.log(5, "Making non-absolute path '%s' be absolute",
-                      maybe_relpath)
-        # Set path to an absolute path, relative to project config.
-        config_dirpath = _os.path.dirname(self.config_file)
-        _LOGGER.log(5, "config_dirpath: %s", config_dirpath)
-        abs_path = _os.path.join(config_dirpath, maybe_relpath)
-        return abs_path
-
-
-    def _handle_missing_env_attrs(self, env_settings_file, when_missing):
-        """ Default environment settings aren't required; warn, though. """
-        missing_env_attrs = \
-            [attr for attr in ["environment", "environment_file"]
-             if not hasattr(self, attr) or getattr(self, attr) is None]
-        if not missing_env_attrs:
-            return
-        message = "'{}' lacks environment attributes: {}".\
-                format(env_settings_file, missing_env_attrs)
-        if when_missing is None:
-            _LOGGER.warn(message)
-        else:
-            when_missing(message)
-
-
-
-@copy
-class Sample(AttributeDict):
-    """
-    Class to model Samples based on a pandas Series.
-
-    :param series: Sample's data.
-    :type series: Mapping | pandas.core.series.Series
-
-    :Example:
-
-    .. code-block:: python
-
-        from models import Project, SampleSheet, Sample
-        prj = Project("ngs")
-        sheet = SampleSheet("~/projects/example/sheet.csv", prj)
-        s1 = Sample(sheet.iloc[0])
-    """
-
-    _FEATURE_ATTR_NAMES = ["read_length", "read_type", "paired"]
-
-    # Originally, this object was inheriting from _pd.Series,
-    # but complications with serializing and code maintenance
-    # made me go back and implement it as a top-level object
-    def __init__(self, series, prj=None):
-
-        # Create data, handling library/protocol.
-        data = dict(series)
-        try:
-            protocol = data.pop("library")
-        except KeyError:
-            pass
-        else:
-            data["protocol"] = protocol
-        super(Sample, self).__init__(entries=data)
-
-        self.prj = prj
-        self.merged_cols = {}
-        self.derived_cols_done = []
-
-        if isinstance(series, _pd.Series):
-            series = series.to_dict()
-        elif isinstance(series, Sample):
-            series = series.as_series().to_dict()
-
-        # Keep a list of attributes that came from the sample sheet,
-        # so we can create a minimal, ordered representation of the original.
-        # This allows summarization of the sample (i.e.,
-        # appending new columns onto the original table)
-        self.sheet_attributes = series.keys()
-
-        # Ensure Project reference is actual Project or AttributeDict.
-        if not isinstance(self.prj, Project):
-            self.prj = AttributeDict(self.prj or dict())
-
-        # Check if required attributes exist and are not empty.
-        missing_attributes_message = self.check_valid()
-        if missing_attributes_message:
-            raise ValueError(missing_attributes_message)
-
-        # Short hand for getting sample_name
-        self.name = self.sample_name
-
-        # Default to no required paths and no YAML file.
-        self.required_paths = None
-        self.yaml_file = None
-
-        # Not yet merged, potentially toggled when merge step is considered.
-        self.merged = False
-
-        # Collect sample-specific filepaths.
-        # Only when sample is added to project, can paths be added.
-        # Essentially, this provides an empty container for tool-specific
-        # filepaths, into which a pipeline may deposit such filepaths as
-        # desired. Such use provides a sort of communication interface
-        # between times and perhaps individuals (processing time vs.
-        # analysis time, and a pipeline author vs. a pipeline user).
-        self.paths = Paths()
-
-
-    def __eq__(self, other):
-        return self.__dict__ == other.__dict__
-
-
-    def __ne__(self, other):
-        return not self == other
-
-
-    def __str__(self):
-        return "Sample '{}'".format(self.name)
-
-
-    @property
-    def input_file_paths(self):
-        """
-        List the sample's data source / input files
-
-        :return list[str]: paths to data sources / input file for this Sample.
-        """
-        return self.data_source.split(" ") if self.data_source else []
-
-
-    def as_series(self):
-        """
-        Returns a `pandas.Series` object with all the sample's attributes.
-
-        :return pandas.core.series.Series: pandas Series representation
-            of this Sample, with its attributes.
-        """
-        # Note that this preserves metadata, but it could be excluded
-        # with self.items() rather than self.__dict__.
-        return _pd.Series(self.__dict__)
-
-
-    def check_valid(self, required=None):
-        """
-        Check provided sample annotation is valid.
-
-        :param Iterable[str] required: collection of required sample attribute
-            names, optional; if unspecified, only a name is required.
-        :return (Exception | NoneType, str, str): exception and messages about
-            what's missing/empty; null with empty messages if there was nothing
-            exceptional or required inputs are absent or not set
-        """
-        missing, empty = [], []
-        for attr in (required or [SAMPLE_NAME_COLNAME]):
-            if not hasattr(self, attr):
-                missing.append(attr)
-            if attr == "nan":
-                empty.append(attr)
-        missing_attributes_message = \
-                "Sample lacks attribute(s). missing={}; empty={}".\
-                        format(missing, empty) if (missing or empty) else ""
-        return missing_attributes_message
-
-
-    def determine_missing_requirements(self):
-        """
-        Determine which of this Sample's required attributes/files are missing.
-
-        :return (type, str): hypothetical exception type along with message
-            about what's missing; null and empty if nothing exceptional
-            is detected
-        """
-
-        null_return = (None, "", "")
-
-        # set_pipeline_attributes must be run first.
-        if not hasattr(self, "required_inputs"):
-            _LOGGER.warn("You must run set_pipeline_attributes "
-                         "before determine_missing_requirements")
-            return null_return
-
-        if not self.required_inputs:
-            _LOGGER.debug("No required inputs")
-            return null_return
-
-        # First, attributes
-        missing, empty = [], []
-        for file_attribute in self.required_inputs_attr:
-            _LOGGER.log(5, "Checking '{}'".format(file_attribute))
-            try:
-                attval = getattr(self, file_attribute)
-            except AttributeError:
-                _LOGGER.log(5, "Missing required input attribute '%s'",
-                             file_attribute)
-                missing.append(file_attribute)
-                continue
-            if attval == "":
-                _LOGGER.log(5, "Empty required input attribute '%s'",
-                             file_attribute)
-                empty.append(file_attribute)
-            else:
-                _LOGGER.log(5, "'{}' is valid: '{}'".
-                              format(file_attribute, attval))
-
-        if missing or empty:
-            reason_key = "Missing and/or empty attribute(s)"
-            reason_detail = "(missing) {}; (empty) {}".format(
-                    ", ".join(missing), ", ".join(empty))
-            return AttributeError, reason_key, reason_detail
-
-        # Second, files
-        missing_files = []
-        for paths in self.required_inputs:
-            _LOGGER.log(5, "Text to split and check paths: '%s'", paths)
-            # There can be multiple, space-separated values here.
-            for path in paths.split(" "):
-                _LOGGER.log(5, "Checking path: '{}'".format(path))
-                if not _os.path.exists(path):
-                    _LOGGER.log(5, "Missing required input file: '{}'".
-                                  format(path))
-                    missing_files.append(path)
-
-        if not missing_files:
-            return null_return
-        else:
-            reason_key = "Missing file(s)"
-            reason_detail = ", ".join(missing_files)
-            return IOError, reason_key, reason_detail
-
-
-    def generate_filename(self, delimiter="_"):
-        """
-        Create a name for file in which to represent this Sample.
-
-        This uses knowledge of the instance's subtype, sandwiching a delimiter
-        between the name of this Sample and the name of the subtype before the
-        extension. If the instance is a base Sample type, then the filename
-        is simply the sample name with an extension.
-
-        :param str delimiter: what to place between sample name and name of
-            subtype; this is only relevant if the instance is of a subclass
-        :return str: name for file with which to represent this Sample on disk
-        """
-        base = self.name if type(self) is Sample else \
-            "{}{}{}".format(self.name, delimiter, self.__class__.__name__)
-        return "{}.yaml".format(base)
-
-
-    def generate_name(self):
-        """
-        Generate name for the sample by joining some of its attribute strings.
-        """
-        raise NotImplementedError("Not implemented in new code base.")
-
-
-    def get_attr_values(self, attrlist):
-        """
-        Get value corresponding to each given attribute.
-
-        :param str attrlist: name of an attribute storing a list of attr names
-        :return list | NoneType: value (or empty string) corresponding to
-            each named attribute; null if this Sample's value for the
-            attribute given by the argument to the "attrlist" parameter is
-            empty/null, or if this Sample lacks the indicated attribute
-        """
-        # If attribute is None, then value is also None.
-        attribute_list = getattr(self, attrlist, None)
-        if not attribute_list:
-            return None
-
-        if not isinstance(attribute_list, list):
-            attribute_list = [attribute_list]
-
-        # Strings contained here are appended later so shouldn't be null.
-        return [getattr(self, attr, "") for attr in attribute_list]
-
-
-    def get_sheet_dict(self):
-        """
-        Create a K-V pairs for items originally passed in via the sample sheet.
-
-        This is useful for summarizing; it provides a representation of the
-        sample that excludes things like config files and derived entries.
-
-        :return OrderedDict: mapping from name to value for data elements
-            originally provided via the sample sheet (i.e., the a map-like
-            representation of the instance, excluding derived items)
-        """
-        return _OrderedDict([[k, getattr(self, k)]
-                             for k in self.sheet_attributes])
-
-
-    def infer_columns(self, implications):
-        """
-        Infer value for additional field(s) from other field(s).
-
-        Add columns/fields to the sample based on values in those already-set
-        that the sample's project defines as indicative of implications for
-        additional data elements for the sample.
-
-        :param Mapping implications: Project's implied columns data
-        :return None: this function mutates state and is strictly for effect
-        """
-
-        _LOGGER.log(5, "Sample attribute implications: {}".
-                    format(implications))
-        if not implications:
-            return
-
-        for implier_name, implied in implications.items():
-            _LOGGER.debug(
-                "Setting Sample variable(s) implied by '%s'", implier_name)
-            try:
-                implier_value = self[implier_name]
-            except KeyError:
-                _LOGGER.debug("No '%s' for this sample", implier_name)
-                continue
-            try:
-                implied_value_by_column = implied[implier_value]
-                _LOGGER.debug("Implications for '%s' = %s: %s",
-                              implier_name, implier_value,
-                              str(implied_value_by_column))
-                for colname, implied_value in \
-                        implied_value_by_column.items():
-                    _LOGGER.log(5, "Setting '%s'=%s",
-                                colname, implied_value)
-                    setattr(self, colname, implied_value)
-            except KeyError:
-                _LOGGER.log(
-                    5, "Unknown implied value for implier '%s' = '%s'",
-                    implier_name, implier_value)
-
-
-    def is_dormant(self):
-        """
-        Determine whether this Sample is inactive.
-
-        By default, a Sample is regarded as active. That is, if it lacks an
-        indication about activation status, it's assumed to be active. If,
-        however, and there's an indication of such status, it must be '1'
-        in order to be considered switched 'on.'
-
-        :return bool: whether this Sample's been designated as dormant
-        """
-        try:
-            flag = self[SAMPLE_EXECUTION_TOGGLE]
-        except KeyError:
-            # Regard default Sample state as active.
-            return False
-        # If specified, the activation flag must be set to '1'.
-        return flag != "1"
-
-
-    @property
-    def library(self):
-        """
-        Backwards-compatible alias.
-
-        :return str: The protocol / NGS library name for this Sample.
-        """
-        warnings.warn("Sample 'library' attribute is deprecated; instead, "
-                      "refer to 'protocol'", DeprecationWarning)
-        return self.protocol
-
-
-    def locate_data_source(self, data_sources, column_name=DATA_SOURCE_COLNAME,
-                           source_key=None, extra_vars=None):
-        """
-        Uses the template path provided in the project config section
-        "data_sources" to piece together an actual path by substituting
-        variables (encoded by "{variable}"") with sample attributes.
-
-        :param Mapping data_sources: mapping from key name (as a value in
-            a cell of a tabular data structure) to, e.g., filepath
-        :param str column_name: Name of sample attribute
-            (equivalently, sample sheet column) specifying a derived column.
-        :param str source_key: The key of the data_source,
-            used to index into the project config data_sources section.
-            By default, the source key will be taken as the value of
-            the specified column (as a sample attribute).
-            For cases where the sample doesn't have this attribute yet
-            (e.g. in a merge table), you must specify the source key.
-        :param dict extra_vars: By default, this will look to
-            populate the template location using attributes found in the
-            current sample; however, you may also provide a dict of extra
-            variables that can also be used for variable replacement.
-            These extra variables are given a higher priority.
-        :return str: regex expansion of data source specified in configuration,
-            with variable substitutions made
-        :raises ValueError: if argument to data_sources parameter is null/empty
-        """
-
-        if not data_sources:
-            return None
-
-        if not source_key:
-            try:
-                source_key = getattr(self, column_name)
-            except AttributeError:
-                reason = "'{attr}': to locate sample's data source, provide " \
-                         "the name of a key from '{sources}' or ensure " \
-                         "sample has attribute '{attr}'".format(
-                         attr=column_name, sources=DATA_SOURCES_SECTION)
-                raise AttributeError(reason)
-
-        try:
-            regex = data_sources[source_key]
-        except KeyError:
-            _LOGGER.debug(
-                    "{}: config lacks entry for data_source key: '{}' "
-                    "in column '{}'; known: {}".format(
-                    self.name, source_key, column_name, data_sources.keys()))
-            return ""
-
-        # Populate any environment variables like $VAR with os.environ["VAR"]
-        regex = _os.path.expandvars(regex)
-
-        try:
-            # Grab a temporary dictionary of sample attributes and update these
-            # with any provided extra variables to use in the replacement.
-            # This is necessary for derived_columns in the merge table.
-            # Here the copy() prevents the actual sample from being
-            # updated by update().
-            temp_dict = self.__dict__.copy()
-            temp_dict.update(extra_vars or dict())
-            val = regex.format(**temp_dict)
-            if '*' in val or '[' in val:
-                _LOGGER.debug("Pre-glob: %s", val)
-                val_globbed = sorted(glob.glob(val))
-                val = " ".join(val_globbed) or val
-                _LOGGER.debug("Post-glob: %s", val)
-
-        except Exception as e:
-            _LOGGER.error("Can't format data source correctly: %s", regex)
-            _LOGGER.error(str(type(e).__name__) + str(e))
-            return regex
-
-        return val
-
-
-    def make_sample_dirs(self):
-        """
-        Creates sample directory structure if it doesn't exist.
-        """
-        for path in self.paths:
-            if not _os.path.exists(path):
-                _os.makedirs(path)
-
-
-    def set_file_paths(self, project=None):
-        """
-        Sets the paths of all files for this sample.
-
-        :param AttributeDict project: object with pointers to data paths and
-            such, either full Project or AttributeDict with sufficient data
-        """
-        # Any columns specified as "derived" will be constructed
-        # based on regex in the "data_sources" section of project config.
-
-        project = project or self.prj
-
-        self.infer_columns(implications=project.get("implied_columns"))
-
-        for col in project.get("derived_columns", []):
-            # Only proceed if the specified column exists
-            # and was not already merged or derived.
-            if not hasattr(self, col):
-                _LOGGER.debug("%s lacks attribute '%s'", self.name, col)
-                continue
-            elif col in self.merged_cols:
-                _LOGGER.debug("'%s' is already merged for %s", col, self.name)
-                continue
-            elif col in self.derived_cols_done:
-                _LOGGER.debug("'%s' has been derived for %s", col, self.name)
-                continue
-            _LOGGER.debug("Deriving column for %s '%s': '%s'",
-                          self.__class__.__name__, self.name, col)
-
-            # Set a variable called {col}_key, so the
-            # original source can also be retrieved.
-            col_key = col + COL_KEY_SUFFIX
-            col_key_val = getattr(self, col)
-            _LOGGER.debug("Setting '%s' to '%s'", col_key, col_key_val)
-            setattr(self, col_key, col_key_val)
-
-            # Determine the filepath for the current data source and set that
-            # attribute on this sample if it's non-empty/null.
-            filepath = self.locate_data_source(
-                    data_sources=project.get(DATA_SOURCES_SECTION),
-                    column_name=col)
-            if filepath:
-                _LOGGER.debug("Setting '%s' to '%s'", col, filepath)
-                setattr(self, col, filepath)
-            else:
-                _LOGGER.debug("Not setting null/empty value for data source "
-                              "'{}': {}".format(col, type(filepath)))
-
-            self.derived_cols_done.append(col)
-
-        # Parent
-        self.results_subdir = project.metadata.results_subdir
-        self.paths.sample_root = sample_folder(project, self)
-
-        # Track url
-        bigwig_filename = self.name + ".bigWig"
-        try:
-            # Project's public_html folder
-            self.bigwig = _os.path.join(
-                    project.trackhubs.trackhub_dir, bigwig_filename)
-            self.track_url = \
-                    "{}/{}".format(project.trackhubs.url, bigwig_filename)
-        except:
-            _LOGGER.debug("No trackhub/URL")
-            pass
-
-    
-    def set_genome(self, genomes):
-        """
-        Set the genome for this Sample.
-
-        :param Mapping[str, str] genomes: genome assembly by organism name
-        """
-        self._set_assembly("genome", genomes)
-        
-        
-    def set_transcriptome(self, transcriptomes):
-        """
-        Set the transcriptome for this Sample.
-
-        :param Mapping[str, str] transcriptomes: transcriptome assembly by
-            organism name
-        """
-        self._set_assembly("transcriptome", transcriptomes)
-        
-        
-    def _set_assembly(self, ome, assemblies):
-        if not assemblies:
-            _LOGGER.debug("Empty/null assemblies mapping")
-            return
-        try:
-            assembly = assemblies[self.organism]
-        except AttributeError:
-            _LOGGER.debug("Sample '%s' lacks organism attribute", self.name)
-            assembly = None
-        except KeyError:
-            _LOGGER.log(5, "Unknown {} value: '{}'".
-                    format(ome, self.organism))
-            assembly = None
-        _LOGGER.log(5, "Setting {} as {} on sample: '{}'".
-                format(assembly, ome, self.name))
-        setattr(self, ome, assembly)
-        
-
-    def set_pipeline_attributes(
-            self, pipeline_interface, pipeline_name, permissive=True):
-        """
-        Set pipeline-specific sample attributes.
-        
-        Some sample attributes are relative to a particular pipeline run,
-        like which files should be considered inputs, what is the total
-        input file size for the sample, etc. This function sets these
-        pipeline-specific sample attributes, provided via a PipelineInterface
-        object and the name of a pipeline to select from that interface.
-        
-        :param PipelineInterface pipeline_interface: A PipelineInterface
-            object that has the settings for this given pipeline.
-        :param str pipeline_name: Which pipeline to choose.
-        :param bool permissive: whether to simply log a warning or error 
-            message rather than raising an exception if sample file is not 
-            found or otherwise cannot be read, default True
-        """
-
-        # Settings ending in _attr are lists of attribute keys.
-        # These attributes are then queried to populate values
-        # for the primary entries.
-        req_attr_names = [("ngs_input_files", "ngs_inputs_attr"),
-                          ("required_input_files", REQUIRED_INPUTS_ATTR_NAME),
-                          ("all_input_files", ALL_INPUTS_ATTR_NAME)]
-        for name_src_attr, name_dst_attr in req_attr_names:
-            _LOGGER.log(5, "Value of '%s' will be assigned to '%s'",
-                        name_src_attr, name_dst_attr)
-            value = pipeline_interface.get_attribute(
-                    pipeline_name, name_src_attr)
-            _LOGGER.log(5, "Assigning '{}': {}".format(name_dst_attr, value))
-            setattr(self, name_dst_attr, value)
-
-        # Post-processing of input attribute assignments.
-        # Ensure that there's a valid all_inputs_attr.
-        if not getattr(self, ALL_INPUTS_ATTR_NAME):
-            required_inputs = getattr(self, REQUIRED_INPUTS_ATTR_NAME)
-            setattr(self, ALL_INPUTS_ATTR_NAME, required_inputs)
-        # Convert attribute keys into values.
-        if self.ngs_inputs_attr:
-            _LOGGER.log(5, "Handling NGS input attributes: '%s'", self.name)
-            # NGS data inputs exit, so we can add attributes like
-            # read_type, read_length, paired.
-            self.ngs_inputs = self.get_attr_values("ngs_inputs_attr")
-
-            set_rtype = False
-            if not hasattr(self, "read_type"):
-                set_rtype_reason = "read_type not yet set"
-                set_rtype = True
-            elif not self.read_type or self.read_type.lower() \
-                    not in VALID_READ_TYPES:
-                set_rtype_reason = "current read_type is invalid: '{}'".\
-                        format(self.read_type)
-                set_rtype = True
-            if set_rtype:
-                _LOGGER.debug(
-                        "Setting read_type for %s '%s': %s",
-                        self.__class__.__name__, self.name, set_rtype_reason)
-                self.set_read_type(permissive=permissive)
-            else:
-                _LOGGER.debug("read_type is already valid: '%s'", 
-                              self.read_type)
-        else:
-            _LOGGER.log(5, "No NGS inputs: '%s'", self.name)
-
-        # Assign values for actual inputs attributes.
-        self.required_inputs = self.get_attr_values(REQUIRED_INPUTS_ATTR_NAME)
-        self.all_inputs = self.get_attr_values(ALL_INPUTS_ATTR_NAME)
-        _LOGGER.debug("All '{}' inputs: {}".format(self.name, self.all_inputs))
-        self.input_file_size = get_file_size(self.all_inputs)
-
-
-    def set_read_type(self, rlen_sample_size=10, permissive=True):
-        """
-        For a sample with attr `ngs_inputs` set, this sets the 
-        read type (single, paired) and read length of an input file.
-
-        :param rlen_sample_size: Number of reads to sample to infer read type,
-            default 10.
-        :type rlen_sample_size: int
-        :param permissive: whether to simply log a warning or error message 
-            rather than raising an exception if sample file is not found or 
-            otherwise cannot be read, default True.
-        :type permissive: bool
-        """
-
-        # TODO: determine how return is being used and standardized (null vs. bool)
-
-        # Initialize the parameters in case there is no input_file, so these
-        # attributes at least exist - as long as they are not already set!
-        for attr in ["read_length", "read_type", "paired"]:
-            if not hasattr(self, attr):
-                _LOGGER.log(5, "Setting null for missing attribute: '%s'", attr)
-                setattr(self, attr, None)
-
-        # ngs_inputs must be set
-        if not self.ngs_inputs:
-            return False
-
-        ngs_paths = " ".join(self.ngs_inputs)
-
-        # Determine extant/missing filepaths.
-        existing_files = list()
-        missing_files = list()
-        for path in ngs_paths.split(" "):
-            if not _os.path.exists(path):
-                missing_files.append(path)
-            else:
-                existing_files.append(path)
-        _LOGGER.debug("{} extant file(s): {}".
-                      format(len(existing_files), existing_files))
-        _LOGGER.debug("{} missing file(s): {}".
-                      format(len(missing_files), missing_files))
-
-        # For samples with multiple original BAM files, check all.
-        files = list()
-        check_by_ftype = {"bam": check_bam, "fastq": check_fastq}
-        for input_file in existing_files:
-            try:
-                file_type = parse_ftype(input_file)
-                read_lengths, paired = check_by_ftype[file_type](
-                        input_file, rlen_sample_size)
-            except (KeyError, TypeError):
-                message = "Input file type should be one of: {}".format(
-                        check_by_ftype.keys())
-                if not permissive:
-                    raise TypeError(message)
-                _LOGGER.error(message)
-                return
-            except NotImplementedError as e:
-                if not permissive:
-                    raise
-                _LOGGER.warn(e.message)
-                return
-            except IOError:
-                if not permissive:
-                    raise
-                _LOGGER.error("Input file does not exist or "
-                              "cannot be read: %s", str(input_file))
-                for feat_name in self._FEATURE_ATTR_NAMES:
-                    if not hasattr(self, feat_name):
-                        setattr(self, feat_name, None)
-                return
-            except OSError as e:
-                _LOGGER.error(str(e) + " [file: {}]".format(input_file))
-                for feat_name in self._FEATURE_ATTR_NAMES:
-                    if not hasattr(self, feat_name):
-                        setattr(self, feat_name, None)
-                return
-
-            # Determine most frequent read length among sample.
-            rlen, _ = sorted(read_lengths.items(), key=itemgetter(1))[-1]
-            _LOGGER.log(5,
-                    "Selected {} as most frequent read length from "
-                    "sample read length distribution: {}".format(
-                            rlen, read_lengths))
-
-            # Decision about paired-end status is majority-rule.
-            if paired > (rlen_sample_size / 2):
-                read_type = "paired"
-                paired = True
-            else:
-                read_type = "single"
-                paired = False
-
-            files.append([rlen, read_type, paired])
-
-        # Check agreement between different files
-        # if all values are equal, set to that value;
-        # if not, set to None and warn the user about the inconsistency
-        for i, feature in enumerate(self._FEATURE_ATTR_NAMES):
-            feature_values = set(f[i] for f in files)
-            if 1 == len(feature_values):
-                feat_val = files[0][i]
-            else:
-                _LOGGER.log(5, "%d values among %d files for feature '%s'",
-                            len(feature_values), len(files), feature)
-                feat_val = None
-            _LOGGER.log(5, "Setting '%s' on %s to %s",
-                        feature, self.__class__.__name__, feat_val)
-            setattr(self, feature, feat_val)
-
-            if getattr(self, feature) is None and len(existing_files) > 0:
-                _LOGGER.warn("Not all input files agree on '%s': '%s'",
-                             feature, self.name)
-
-
-    def to_yaml(self, path=None, subs_folder_path=None, delimiter="_"):
-        """
-        Serializes itself in YAML format.
-
-        :param str path: A file path to write yaml to; provide this or
-            the subs_folder_path
-        :param str subs_folder_path: path to folder in which to place file
-            that's being written; provide this or a full filepath
-        :param str delimiter: text to place between the sample name and the
-            suffix within the filename; irrelevant if there's no suffix
-        :return str: filepath used (same as input if given, otherwise the
-            path value that was inferred)
-        :raises ValueError: if neither full filepath nor path to extant
-            parent directory is provided.
-        """
-
-        # Determine filepath, prioritizing anything given, then falling
-        # back to a default using this Sample's Project's submission_subdir.
-        # Use the sample name and YAML extension as the file name,
-        # interjecting a pipeline name as a subfolder within the Project's
-        # submission_subdir if such a pipeline name is provided.
-        if not path:
-            if not subs_folder_path:
-                raise ValueError(
-                    "To represent {} on disk, provide a full path or a path "
-                    "to a parent (submissions) folder".
-                    format(self.__class__.__name__))
-            _LOGGER.debug("Creating filename for %s: '%s'",
-                          self.__class__.__name__, self.name)
-            filename = self.generate_filename(delimiter=delimiter)
-            _LOGGER.debug("Filename: '%s'", filename)
-            path = _os.path.join(subs_folder_path, filename)
-
-        _LOGGER.debug("Setting %s filepath: '%s'",
-                      self.__class__.__name__, path)
-        self.yaml_file = path
-
-        def _is_project(obj, name=None):
-            """ Determine if item to prep for disk is Sample's project. """
-            return name == "prj"
-
-        def obj2dict(obj, name=None,
-                to_skip=("merge_table", "samples", "sheet", "sheet_attributes")):
-            """
-            Build representation of object as a dict, recursively
-            for all objects that might be attributes of self.
-
-            :param object obj: what to serialize to write to YAML.
-            :param str name: name of the object to represent.
-            :param Iterable[str] to_skip: names of attributes to ignore.
-            """
-            if name:
-                _LOGGER.log(5, "Converting to dict: '{}'".format(name))
-            if _is_project(obj, name):
-                _LOGGER.debug("Attempting to store %s's project metadata",
-                              self.__class__.__name__)
-                prj_data = grab_project_data(obj)
-                _LOGGER.debug("Sample's project data: {}".format(prj_data))
-                return {k: obj2dict(v, name=k) for k, v in prj_data.items()}
-            if isinstance(obj, list):
-                return [obj2dict(i) for i in obj]
-            if isinstance(obj, AttributeDict):
-                return {k: obj2dict(v, name=k) for k, v in obj.__dict__.items()
-                        if k not in to_skip and
-                        (k not in ATTRDICT_METADATA or
-                         v != ATTRDICT_METADATA[k])}
-            elif isinstance(obj, Mapping):
-                return {k: obj2dict(v, name=k)
-                        for k, v in obj.items() if k not in to_skip}
-            elif isinstance(obj, (Paths, Sample)):
-                return {k: obj2dict(v, name=k)
-                        for k, v in obj.__dict__.items() if
-                        k not in to_skip}
-            elif isinstance(obj, _pd.Series):
-                _LOGGER.warn("Serializing series as mapping, not array-like")
-                return obj.to_dict()
-            elif hasattr(obj, 'dtype'):  # numpy data types
-                # TODO: this fails with ValueError for multi-element array.
-                return obj.item()
-            elif _pd.isnull(obj):
-                # Missing values as evaluated by pd.isnull().
-                # This gets correctly written into yaml.
-                return "NaN"
-            else:
-                return obj
-
-        _LOGGER.debug("Serializing %s: '%s'",
-                      self.__class__.__name__, self.name)
-        serial = obj2dict(self)
-
-        # TODO: this is the way to add the project metadata reference if
-        # the metadata items are to be accessed directly on the Sample rather
-        # than through the Project; that is:
-        #
-        # sample.output_dir
-        # instead of
-        # sample.prj.output_dir
-        #
-        # In this case, "prj" should be added to the default argument to the
-        # to_skip parameter in the function signature, and the instance check
-        # of the object to serialize against Project can be dropped.
-        """
-        try:
-            serial.update(self.prj.metadata)
-        except AttributeError:
-            _LOGGER.debug("%s lacks %s reference",
-                          self.__class__.__name__, Project.__class__.__name__)
-        else:
-            _LOGGER.debug("Added %s metadata to serialized %s",
-                          Project.__class__.__name__, self.__class__.__name__)
-        """
-
-        with open(self.yaml_file, 'w') as outfile:
-            _LOGGER.debug("Generating YAML data for %s: '%s'",
-                          self.__class__.__name__, self.name)
-            try:
-                yaml_data = yaml.safe_dump(serial, default_flow_style=False)
-            except yaml.representer.RepresenterError:
-                _LOGGER.error("SERIALIZED SAMPLE DATA: {}".format(serial))
-                raise
-            outfile.write(yaml_data)
-
-
-    def update(self, newdata):
-        """
-        Update Sample object with attributes from a dict.
-        """
-        for key, value in newdata.items():
-            setattr(self, key, value)
 
 
 
@@ -3122,29 +828,6 @@ class _InvalidResourceSpecificationException(Exception):
     """ Pipeline interface resources--if present--needs default. """
     def __init__(self, reason):
         super(_InvalidResourceSpecificationException, self).__init__(reason)
-
-
-
-class _MetadataOperationException(Exception):
-    """ Illegal/unsupported operation, motivated by `AttributeDict`. """
-
-    def __init__(self, obj, meta_item):
-        """
-        Instance with which the access attempt was made, along with the
-        name of the reserved/privileged metadata item, define the exception.
-
-        :param object obj: instance with which
-            offending operation was attempted
-        :param str meta_item: name of the reserved metadata item
-        """
-        try:
-            classname = obj.__class__.__name__
-        except AttributeError:
-            # Maybe we were given a class or function not an instance?
-            classname = obj.__name__
-        explanation = "Attempted unsupported operation on {} item '{}'". \
-                format(classname, meta_item)
-        super(_MetadataOperationException, self).__init__(explanation)
